Add files via upload
Swapped the "RGB" and "Monochrome" on the home page to correlate with their specific pages better. Added a smaller logo to the RBG and Monochrome pages. Decreased the size of the canvas so that the audio is able to be interacted with again (changing time, pausing, and volume). Increased the bar size due to the decrease in size after changing the canvas height. Added an underline to the "Home" button.